perf: reduce edge bundle size for middleware and dashboard overlays
