+++ app/auth/layout.tsx
@@ -2,6 +2,7 @@ import type React from "react"
 import type { Metadata } from "next"
 import { headers } from "next/headers"
 import { CSPNonceSync } from "@/components/providers/csp-nonce-sync"
+import { redirectAuthenticatedAuthRoute } from "@/lib/server/route-access"
 
 // Cloudflare Pages requires dynamic routes to be marked as edge
 export const runtime = 'edge'
@@ -21,6 +22,7 @@ export default async function AuthLayout({
 }: {
   children: React.ReactNode
 }) {
+  await redirectAuthenticatedAuthRoute()
   const nonce = (await headers()).get('x-nonce')
 
   return (
